Refactor: add register state type and re-packing
